% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportedUtils.R
\name{checkForFcsByteOffsetIssue}
\alias{checkForFcsByteOffsetIssue}
\title{Check for byte offset corruption in the header of an .fcs file}
\usage{
checkForFcsByteOffsetIssue(fcsFile)
}
\arguments{
\item{fcsFile}{A string containing the location of an .fcs file}
}
\value{
Boolean: TRUE if there's an offset issue detected, FALSE otherwise
}
\description{
Opens an .fcs file in binary mode and reads the 2 values of
the byte offset from the text header. These should agree with one another as
a check for file data integrity, but in practice the values are often off
by one. This seems to be related to exporting rare gated populations from
flowJo, and can often be corrected by re-gating and re-exporting.
}
\author{
Mario G Rosasco, \email{mrosasco@benaroyaresearch.org}
}
