% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportedUtils.R
\name{downsampleFcsList}
\alias{downsampleFcsList}
\title{Downsample a set of FCS files reproducibly, to make the event count more uniform and reduce memory usage}
\usage{
downsampleFcsList(
  fcsInfoFile,
  popsToDownsample = "all",
  maxEvents = 1e+05,
  downsampleMode = "storeVectors",
  outfileFcsInfoDownsampled,
  outdirFcsFilesDownsampled,
  seed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{fcsInfoFile}{a character string indicating the path to a file with
information on FCS files to be used.}

\item{popsToDownsample}{a character vector with the names of the cell
populations to be downsampled. Allows the user to keep all events from
certain populations even if the number of events exceed \code{maxEvents}, 
then \code{maxEvents} rows are selected from the FCS file for future use.
For example, users may want to keep all events from antigen-specific
populations, while downsampling the parent populations.}

\item{maxEvents}{numeric, the maximum number of events to be kept for each
FCS file. This cap is applied only to files matching populations in
\code{popsToDownsample}.}

\item{downsampleMode}{(default: "storeVectors") character, either 
"storeVectors" or "writeFiles" (or unique abbreviations thereof). For 
"storeVectors", a list of vectors is returned, with one element for each file
listed in \code{fcsInfoFile}; this should be passed to 
\code{setupDiscovrExperiment} as \code{downSampleRowList}. For "writeFiles", 
you must specify a valid directory in \code{outdirFcsFilesDownsampled} to 
write new FCS files to, and a valid path for \code{outfileFcsInfoDownsampled}.}

\item{outfileFcsInfoDownsampled}{character string, the location to write out 
the FCS info file with updated paths for the downsampled files.  Used only if
\code{downsampleMode == "writeFiles"}.}

\item{outdirFcsFilesDownsampled}{character string, the path for the directory
to write out the the downsampled files. Used only if 
\code{downsampleMode == "writeFiles"}.}

\item{seed}{(default: NULL) numeric, the seed to be passed to 
\code{set.seed} to make the downsampling reproducible.}

\item{verbose}{(default: TRUE) A logical specifying whether to display processing messages}
}
\value{
Depends on \code{downsampleMode}. If 
\code{downsampleMode == "storeVectors"}, a list of vectors is returned, one 
for each row in the file at fcsInfoFile, with the row numbers of events to 
be included. This object should be passed to \code{setupDiscovrExperiment}
via argument \code{downsampleVectorList} to enforce the downsampling. If 
\code{downsampleMode == "writeFiles"}, invisibly returns \code{NULL}.
}
\description{
This function downsamples a set of FCS cytometry data files to a maximum
event count. For each FCS file representing a population in 
\code{popsToDownsample}, it checks whether the file exceeds the maximum
event count, set by \code{maxEvents}. If the maximum event count is
exceeded, it selects a random set of rows. These rows can either be stored
as a list of vectors and passed to \code{setupDiscovrExperiment}, or new
FCS files can be written out to a specified directory, with a new fcsInfoFile
specifying the updated file paths. Results can be made reproducible by
passing a non-NULL value for \code{seed}.
}
\details{
Files specified in \code{fcsInfoFile} are checked for a number of
potential issues, including missing paths, no events, and byte offset issues.
}
\author{
Matthew J Dufort, \email{mdufort@benaroyaresearch.org}
}
