% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovrExperiment.R
\name{getSampleNames}
\alias{getSampleNames}
\title{Get all sample IDs in an experiment}
\usage{
getSampleNames(x)
}
\arguments{
\item{x}{discovrExperiment object}
}
\description{
Get all sample IDs in an experiment
}
\author{
Mario G Rosasco
}
