% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportedUtils.R
\name{getSubjectClusters}
\alias{getSubjectClusters}
\title{Display the per-subject number of clusters in an experiment}
\usage{
getSubjectClusters(experiment)
}
\arguments{
\item{experiment}{A discovrExperiment created using \code{setupExperiment()}}
}
\value{
A data frame containing the number of events detected for each subject
}
\description{
Display the per-subject number of clusters in an experiment
}
\author{
Mario G Rosasco, \email{mrosasco@benaroyaresearch.org}, Virginia Muir
}
