% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportedUtils.R
\name{getSubjectCounts}
\alias{getSubjectCounts}
\title{Display the per-subject event counts in an experiment}
\usage{
getSubjectCounts(experiment)
}
\arguments{
\item{experiment}{A discovrExperiment created using \code{setupDiscovrExperiment()}}
}
\value{
A data frame containing the number of events detected for each subject
}
\description{
Display the per-subject event counts in an experiment
}
\author{
Mario G Rosasco, \email{mrosasco@benaroyaresearch.org}, Virginia Muir
}
