% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapGenerators.R
\name{makeMetaclusterHeatmaps}
\alias{makeMetaclusterHeatmaps}
\title{Perform metaclustering for a DISCOV-R experiment}
\usage{
makeMetaclusterHeatmaps(
  experiment,
  parentSubset = NA,
  childSubsets = NA,
  filenamePrefix = paste0(format(Sys.Date(), "\%y\%m\%d"), "-Zscores-perSubj"),
  verbose = TRUE
)
}
\arguments{
\item{experiment}{A discovrExperiment created using \code{setupDiscovrExperiment},
clustered using \code{clusterDiscovrExperiment}, and metaclustered with
\code{metaclusterDiscovrExperiment}}

\item{parentSubset}{A string indicating which subset represents the parent population (eg: all CD8+ cells).
If this is left NA, then it's assumed the first subset listed in the experiment's fcsInfo file is the parent
population and all other subsets are rare populations. (default: NA)}

\item{childSubsets}{A vector of strings indicating which subsets represent rare populations (eg: Tmr+ cells).
If this is left NA, then it's assumed the first subset listed in the experiment's fcsInfo file is the parent
population and all other subsets are rare populations. (default: NA)}

\item{filenamePrefix}{A string indicating the prefix to use for each file that gets saved.
This can include a directory path. (default: paste0(format(Sys.Date(), "%y%m%d"), "-Zscores-perSubj"))}

\item{verbose}{A boolean specifying whether to display processing messages (default: TRUE)}
}
\value{
An S3 object of class \code{discovrExperiment}
}
\description{
Perform metaclustering for a DISCOV-R experiment
}
\seealso{
\code{\link{setupDiscovrExperiment}} \code{\link{clusterDiscovrExperiment}} \code{\link{metaclusterDiscovrExperiment}}
}
\author{
Mario G Rosasco, \email{mrosasco@benaroyaresearch.org}, Virginia Muir
}
