% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclusteringPhase.R
\name{metaclusterDiscovrExperiment}
\alias{metaclusterDiscovrExperiment}
\title{Perform metaclustering for a DISCOV-R experiment}
\usage{
metaclusterDiscovrExperiment(
  experiment,
  dropSamples = NA,
  dropMarkers = NA,
  pctInClusterThreshold = 1,
  nMetaclusters = 12,
  linkage = "ward.D2",
  distance = "euclidean",
  verbose = TRUE
)
}
\arguments{
\item{experiment}{A discovrExperiment created using \code{setupDiscovrExperiment}
and clustered using \code{clusterDiscovrExperiment}}

\item{dropSamples}{A string or vector of strings indicating names of samples
to exclude from further analysis. Note that the convenience function
\code{getSampleNames} can be used to view a list of all sample names. (default: NA)}

\item{dropMarkers}{A string or vector of strings indicating names of markers to exclude
from analysis. Note that printing the experiment object will display a list of all
markers. By default all clustering markers will be included. (default: NA)}

\item{pctInClusterThreshold}{A numeric indicating a percentage. Clusters with event
occupancy below this threshold will be considered 'low-abundance' and will be
dropped from further analysis (default: 1)}

\item{nMetaclusters}{A numeric indicating the number of metaclusters to generate (default: 12)}

\item{linkage}{A string indicating the cluster linkage method to use when metaclustering.
See \code{hclust} for more details. (default: "ward.D2")}

\item{distance}{A string indicating the distance metric to use when metaclustering. (default: "euclidean")}

\item{verbose}{A boolean specifying whether to display processing messages (default: TRUE)}
}
\value{
An S3 object of class \code{discovrExperiment}
}
\description{
Perform metaclustering for a DISCOV-R experiment
}
\seealso{
\code{\link{setupDiscovrExperiment}} \code{\link{clusterDiscovrExperiment}}
}
\author{
Mario G Rosasco, Virginia Muir

Matthew J Dufort, \email{mdufort@benaroyaresearch.org}
}
