% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportedUtils.R
\name{plotDensityNormalizedExprsDiscovrExperiment}
\alias{plotDensityNormalizedExprsDiscovrExperiment}
\title{Plot normalized vs. pre-normalization marker level distributions for a discovrExperiment object}
\usage{
plotDensityNormalizedExprsDiscovrExperiment(
  experiment,
  clusteringMarkersOnly = TRUE,
  normalizationInfo = NULL,
  filenameOut = NULL,
  width = 12,
  height = 6,
  ...
)
}
\arguments{
\item{experiment}{A discovrExperiment created using
\code{clusterDiscovrExperiment}, \code{normalizeDiscovrExperiment},
\code{metaclusterDiscovrExperiment}. Must have status "clustered",
"normalized", or "metaclustered". If it has status "clustered", it will be
passed to \code{normalizeDiscovrExperiment} along with any relevant
arguments prior to plotting.}

\item{clusteringMarkersOnly}{(default: NULL)}

\item{normalizationInfo}{(default: NULL) Optional object containing
information on how to normalize, passed to \code{normalizeDiscovrExperiment}
(see documentation of that function for usage details). If this argument
is non-null, normalization will be performed prior to plotting, even if the
'experiment' object has already been normalized. This allows for testing a
range of normalization approaches on a discovrExperiment object. If
\code{experiment} has status "clustered", it must either have a non-null
value for this argument, or must contain normalization methods in the
markerInfo element, per the documentation for
\code{normalizeDiscovrExperiment}.
value must be provided if \code{experiment} has status "clustered" or does
not contain the element \code{mergedExprNormalizedScaled} which is generated
by \code{normalizeDiscovrExperiment}.}

\item{filenameOut}{(default: NULL) Path for outputing a file with plots. If
null, no file is generated. The extension is used to determine plot type,
which can be "pdf", "png", or "svg".}

\item{width}{(default: 12) Numeric, width of output plot device in inches}

\item{height}{(default: 6) Numeric, height of output plot device in inches}

\item{...}{optional arguments passed to \code{normalizeDiscovrExperiment}}
}
\value{
A list of ggplot objects, one for each marker. Plots will be output
to file if filenameOut is specified. To suppress output to the default
plotting device, assign the result to an object, or run the function with
\code{invisible()}
}
\description{
This function generates a set of density plots of the pre- and
post-normalization expression values from a discovrExperiment object. If
the object has status "normalized" or "metaclustered", the normalized values
will be used. If the object has status "clustered", normalization will be
attempted using information provided, passed to
\code{normalizeDiscovrExperiment}. The function returns a list of plots as
ggplot2 objects, and can optionally output plots to a single file containing
a separate page for each marker. The purpose of this function is to evaluate
potential normalization methods for each marker, based on distributions of
the marker levels before and after normalization.
}
\author{
Matthew J Dufort, \email{mdufort@benaroyaresearch.org}
}
