% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovrExperiment.R
\name{print.discovrExperiment}
\alias{print.discovrExperiment}
\title{Print method for discovrExperiment objects}
\usage{
\method{print}{discovrExperiment}(x, ...)
}
\arguments{
\item{x}{discovrExperiment object}

\item{...}{other arguments to be passed to generic print method}
}
\description{
Prints summary information about the experiment structure.
}
\author{
Mario G Rosasco
}
