% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclusteringPhase.R
\name{recutMetaclusters}
\alias{recutMetaclusters}
\title{Re-cut a metaclustered experiment with a different target number of metaclusters}
\usage{
recutMetaclusters(
  experiment,
  nMetaclusters = 12,
  linkage = "ward.D2",
  distance = "euclidean",
  verbose = TRUE
)
}
\arguments{
\item{experiment}{A discovrExperiment created using \code{setupDiscovrExperiment},
clustered using \code{clusterDiscovrExperiment}, and metaclustered using
\code{metaclusterDiscovrExperiment}}

\item{nMetaclusters}{A numeric indicating the number of metaclusters to generate (default: 12)}

\item{linkage}{A string indicating the cluster linkage method to use when metaclustering.
See \code{hclust} for more details. (default: "ward.D2")}

\item{distance}{A string indicating the distance metric to use when metaclustering. (default: "euclidean")}

\item{verbose}{A boolean specifying whether to display processing messages (default: TRUE)}
}
\value{
An S3 object of class \code{discovrExperiment}
}
\description{
Re-cut a metaclustered experiment with a different target number of metaclusters
}
\seealso{
\code{\link{metaclusterDiscovrExperiment}}
}
\author{
Mario G Rosasco, Virginia Muir
}
