% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportedUtils.R
\name{testNormalizationMethodByMarker}
\alias{testNormalizationMethodByMarker}
\title{Test a range of normalization options on a marker in a discovrExperiment}
\usage{
testNormalizationMethodByMarker(
  experiment,
  markers,
  normalizationMethods = c("none", "zScore", "warpSet"),
  seed = 12345
)
}
\arguments{
\item{experiment}{A discovrExperiment created using
\code{clusterDiscovrExperiment}, \code{normalizeDiscovrExperiment}, or
\code{metaclusterDiscovrExperiment}. In order to return metacluster numbers
for each cell, its status must be "metaclustered".}

\item{markers}{a character vector indicating the markers on which to test
normalization}

\item{normalizationMethods}{(default: c("none", "zScore", "warpSet")) a
character vector with the normalization methods to be tested. All values
should be acceptable inputs for the argument 'normalizationMethod' in
\code{normalizeDiscovrExperiment}}

\item{seed}{(default: 12345) Numeric, the random number seed for warpSet
normalization, passed to \code{normalizeWarpSetMergedExpr}}
}
\value{
A reduced \code{discovrExperiment} object (or named list of objects),
each containing only the marker listed, with normalized values in
\code{experiment$mergedExprNormalizedScaled}. If \code{markers} has length >
1, the function will return a named list, with each element containing a
\code{discovrExperiment} object normalized using the given method.
}
\description{
This function tests different normalization methods on a marker in a
\code{discovrExperiment} object. Its purpose is to facilitate selection of
the appropriate normalization method for markers where this is not
immediately apparent. It is currently designed to run on one marker at a
time. The results can be examined for appropriate distributions across
samples, or plotted sequentially using
\code{plotDensityNormalizedExprsDiscovrExperiment}
}
\author{
Matthew J Dufort, \email{mdufort@benaroyaresearch.org}
}
