% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportedUtils.R
\name{viewSubjectCounts}
\alias{viewSubjectCounts}
\title{Display the per-subject event counts in an experiment}
\usage{
viewSubjectCounts(experiment)
}
\arguments{
\item{experiment}{A discovrExperiment created using \code{setupExperiment()}}
}
\description{
Display the per-subject event counts in an experiment
}
\author{
Mario G Rosasco, \email{mrosasco@benaroyaresearch.org}, Virginia Muir
}
