% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportedUtils.R
\name{checkForFcsByteOffsetIssue}
\alias{checkForFcsByteOffsetIssue}
\title{Check for byte offset corruption in the header of an .fcs file}
\usage{
checkForFcsByteOffsetIssue(fcsFile)
}
\arguments{
\item{fcsFile}{A string containing the location of an .fcs file}
}
\value{
Boolean: TRUE if there's an offset issue detected, FALSE otherwise
}
\description{
Uses flowCore functions to check the 2 values of the byte offset from the
text header. These should agree with one another as a check for file data
integrity, but in practice the values are often off by one. This seems to
be related to exporting rare gated populations from flowJo, and can often
be corrected by re-gating and re-exporting. This function previously used
custom code that threw an error when one set of values were 0s; the
flowCore functions have better handling for this case.
}
\author{
Mario G Rosasco

Matthew J Dufort, \email{mdufort@benaroyaresearch.org}
}
