% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringPhase.R
\name{clusterDiscovrExperiment}
\alias{clusterDiscovrExperiment}
\title{Perform phenograph clustering for a DISCOV-R experiment}
\usage{
clusterDiscovrExperiment(experiment, method = "phenograph", verbose = TRUE)
}
\arguments{
\item{experiment}{A discovrExperiment created using \code{setupExperiment()}}

\item{method}{A character string indicating the clustering method to use.
Currently only 'phenograph' is supported as a clustering method.}

\item{verbose}{A boolean specifying whether to display processing messages (default: TRUE)}
}
\value{
An S3 object of class \code{discovrExperiment}
}
\description{
Perform phenograph clustering for a DISCOV-R experiment
}
\seealso{
\code{\link{setupDiscovrExperiment}} \code{\link{setupDiscovrExperiment}}
}
\author{
Mario G Rosasco, \email{mrosasco@benaroyaresearch.org}, Virginia Muir
}
