% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportedUtils.R
\name{getMetaclusterOccupancy}
\alias{getMetaclusterOccupancy}
\title{Display the fraction of events that are found in each metacluster for each
sample and cell subset}
\usage{
getMetaclusterOccupancy(experiment)
}
\arguments{
\item{experiment}{A discovrExperiment created using \code{setupDiscovrExperiment()}}
}
\value{
A data frame containing the fraction of events in each metacluster
}
\description{
Display the fraction of events that are found in each metacluster for each
sample and cell subset
}
\author{
Mario G Rosasco, \email{mrosasco@benaroyaresearch.org}
}
