% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapGenerators.R
\name{makeMetaclusterHeatmaps}
\alias{makeMetaclusterHeatmaps}
\title{Generate heatmaps from a metaclustered experiment}
\usage{
makeMetaclusterHeatmaps(
  experiment,
  childSubsets = NA,
  filenamePrefix = NA,
  parentTitle = "Parent",
  metaclusterColors = NA,
  zScoreBreaks = c(-2, 0, 2),
  zScoreColors = c("#00FFFF", "#000000", "#FDE725"),
  intensityBreaks = c(0, 2, 4, 6),
  intensityColors = c("#4A0E53", "#925F54", "#D6B343", "#FDE727"),
  columnDendHeight = 10,
  rowDendWidth = 10,
  verbose = TRUE
)
}
\arguments{
\item{experiment}{A discovrExperiment created using \code{setupDiscovrExperiment},
clustered using \code{clusterDiscovrExperiment}, and metaclustered with
\code{metaclusterDiscovrExperiment}}

\item{childSubsets}{A vector of strings indicating which subsets represent rare populations (eg: Tmr+ cells).
If this is left NA, then it's assumed all non-parent subsets are rare populations. (default: NA)}

\item{filenamePrefix}{A string indicating the prefix to use for each file that gets saved.
This can include a directory path. If left as NA, will use the current date as the prefix
with the format YYMMDD. (default: NA)}

\item{parentTitle}{A string to use to label the parent population heatmaps (default: "Parent")}

\item{metaclusterColors}{An optional vector of color values to use to label metaclusters in heatmaps.
If left as the default 'NA', a random list of visually distinct colors will be generated.
Note that if a color vector is supplied, it must contain at least as many colors as there are metaclusters. (default: NA)}

\item{zScoreBreaks}{A vector of numbers indicating the values used to calibrate the Z score color scale. This vector
must have the same number of values as \code{zScoreColors}. (default: c(-2,0,2))}

\item{zScoreColors}{A vector of strings indicating the colors used to represent the Z score value range. This vector
must have the same number of values as \code{zScoreBreaks}. (default: c("#00FFFF", "#000000", "#FDE725"))}

\item{intensityBreaks}{A vector of numbers indicating the values used to calibrate the transformed marker intensity
scale. This vector must have the same number of values as \code{intensityColors} (default: c(0,2,4,6))}

\item{intensityColors}{A vector of strings indicating the colors used to represent the transformed marker intensity
value range. This vector must have the same number of values as \code{intensityBreaks}
(default: c("#4A0E53", "#925F54", "#D6B343", "#FDE727"))}

\item{columnDendHeight}{A numeric indicating the column dendrogram height (in mm)
for the Z score cluster phenotypes heatmap (default: 10)}

\item{rowDendWidth}{A numeric indicating the row dendrogram width (in mm)
for the Z score cluster phenotypes heatmap (default: 10)}

\item{verbose}{A boolean specifying whether to display processing messages (default: TRUE)}
}
\value{
An S3 object of class \code{discovrExperiment}
}
\description{
Generate heatmaps from a metaclustered experiment
}
\seealso{
\code{\link{setupDiscovrExperiment}} \code{\link{clusterDiscovrExperiment}} \code{\link{metaclusterDiscovrExperiment}}
}
\author{
Mario G Rosasco, \email{mrosasco@benaroyaresearch.org}, Virginia Muir
}
