% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapGenerators.R
\name{makeMetaclusterHeatmaps}
\alias{makeMetaclusterHeatmaps}
\title{Generate heatmaps from a metaclustered experiment}
\usage{
makeMetaclusterHeatmaps(
  experiment,
  parentSubset = NA,
  childSubsets = NA,
  filenamePrefix = NA,
  parentTitle = "Parent",
  verbose = TRUE
)
}
\arguments{
\item{experiment}{A discovrExperiment created using \code{setupDiscovrExperiment},
clustered using \code{clusterDiscovrExperiment}, and metaclustered with
\code{metaclusterDiscovrExperiment}}

\item{parentSubset}{A string indicating which subset represents the parent population (eg: all CD8+ cells).
If this is left NA, then it's assumed the first subset listed in the experiment's fcsInfo file is the parent
population and all other subsets are rare populations. (default: NA)}

\item{childSubsets}{A vector of strings indicating which subsets represent rare populations (eg: Tmr+ cells).
If this is left NA, then it's assumed the first subset listed in the experiment's fcsInfo file is the parent
population and all other subsets are rare populations. (default: NA)}

\item{filenamePrefix}{A string indicating the prefix to use for each file that gets saved.
This can include a directory path. If left as NA, will use the current date as the prefix
with the format YYMMDD. (default: NA)}

\item{parentTitle}{A string to use to label the parent population heatmaps (default: "Parent")}

\item{verbose}{A boolean specifying whether to display processing messages (default: TRUE)}
}
\value{
An S3 object of class \code{discovrExperiment}
}
\description{
Generate heatmaps from a metaclustered experiment
}
\seealso{
\code{\link{setupDiscovrExperiment}} \code{\link{clusterDiscovrExperiment}} \code{\link{metaclusterDiscovrExperiment}}
}
\author{
Mario G Rosasco, \email{mrosasco@benaroyaresearch.org}, Virginia Muir
}
