% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclusteringPhase.R
\name{metaclusterDiscovrExperiment}
\alias{metaclusterDiscovrExperiment}
\title{Perform metaclustering for a DISCOV-R experiment}
\usage{
metaclusterDiscovrExperiment(
  experiment,
  dropSamples = NA,
  dropMarkers = NA,
  pctInClusterThreshold = 1,
  verbose = TRUE
)
}
\arguments{
\item{experiment}{A discovrExperiment created using \code{setupDiscovrExperiment}
and clustered using \code{clusterDiscovrExperiment}}

\item{dropSamples}{A string or vector of strings indicating names of samples
to exclude from further analysis. Note that the convenience function
\code{getSampleNames} can be used to view a list of all sample names. (default: NA)}

\item{dropMarkers}{A string or vector of strings indicating names of markers to exclude
from analysis. Note that printing the experiment object will display a list of all
markers. By default all markers will be included. (default: NA)
'markerInfoFile'}

\item{pctInClusterThreshold}{A numeric indicating a percentage. Cluster with event
occupancy below this threshold will be considered 'low-abundance' and will be
dropped from further analysis (default: 1)}

\item{verbose}{A boolean specifying whether to display processing messages (default: TRUE)}
}
\value{
An S3 object of class \code{discovrExperiment}
}
\description{
Perform metaclustering for a DISCOV-R experiment
}
\seealso{
\code{\link{setupDiscovrExperiment}} \code{\link{setupDiscovrExperiment}}
}
\author{
Mario G Rosasco, \email{mrosasco@benaroyaresearch.org}, Virginia Muir
}
