% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringPhase.R
\name{normalizeDiscovrExperiment}
\alias{normalizeDiscovrExperiment}
\title{Normalize marker expression within a DISCOV-R experiment}
\usage{
normalizeDiscovrExperiment(
  experiment,
  normalizationInfo = NULL,
  defaultNormalizationMethod = "zScore",
  seed = 12345,
  verbose = TRUE
)
}
\arguments{
\item{experiment}{A discovrExperiment created using
\code{setupDiscovrExperiment()} and clustered using
\code{clusterDiscovrExperiment()}}

\item{normalizationInfo}{(default: NULL) Optional object containing
information on how to normalize. If left NULL, the \code{experiment} object
will be checked for normalization info in
\code{experiment$markerInfo$normalizationMethod}. If that is not present,
or for any markers without a method specified, normalization will use
the default method, currently "zScore". Acceptable non-NULL values include a
data frame with columns columns "commonMarkerName" and "normalizationMethod",
a named vector with elements specifying normalization method and names
referring to each marker in \code{experiment$mergedExpr}, or a single
character value specifying the normalization method to be applied to all
markers. Options for normalization are "zScore", "none", and "warpSet". In
addition, "warpSet" can include a maximum peak number specification as a
number immediately following (e.g. "warpSet2") - see documentation of
\code{flowStats::warpSet} for details. Any markers lacking a specification
will be normalized using the default method (currently "zScore").}

\item{defaultNormalizationMethod}{(default: "zScore") A character string, the
normalization method to use for markers without another method specified.
Options are the same as in \code{normalizationInfo} above.}

\item{seed}{(default: 12345) Numeric, the random number seed for warpSet
normalization, passed to \code{normalizeWarpSetMergedExpr}}

\item{verbose}{(default: TRUE) A logical specifying whether to display processing messages}
}
\value{
An S3 object of class \code{discovrExperiment}, with additional
elements named "mergedExprNormalizedScaled" and
"clusterMeansNormalizedScaled". If the normalization methods were not taken
from \code{experiment$markerInfo$normalizationMethod}, the normalization
method for each marker will be stored there.
}
\description{
Normalize marker expression within a DISCOV-R experiment
}
\seealso{
\code{\link{discovrExperiment}}
}
\author{
Matt Dufort
}
