% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringPhase.R
\name{setupDiscovrExperiment}
\alias{setupDiscovrExperiment}
\title{Load in a dataset and perform phenograph clustering}
\usage{
setupDiscovrExperiment(
  markerInfoFile,
  fcsInfoFile,
  markerCommonField = "fixed",
  markerFcsField = "desc",
  arcsinhA = 0,
  arcsinhB = 0.2,
  arcsinhC = 0,
  verbose = TRUE
)
}
\arguments{
\item{markerInfoFile}{A character string indicating the path to a .csv file. This file is expected to have columns
named useToCluster", as well as the names specified in the \code{markerCommonField} and \code{markerFcsField} variables.
Details of the \code{markerCommonField} and \code{markerFcsField} arguments are provided below; the "useToCluster"
column should have only TRUE or FALSE values. Markers with a TRUE value in this column will be used for clustering,
whereas the others will not.}

\item{fcsInfoFile}{A character string indicating the path to a file containing columns named "subject",
"cellSubset", and "filename". The "filename" field must contain paths to the .fcs files that will be used in analysis.}

\item{markerCommonField}{A character string indicating the name of a column containing common marker names for human use,
like "CD45" (default: "fixed")}

\item{markerFcsField}{A character string indicating the name of a column containing the marker names in the .fcs files,
like "89Y_CD45" (default: "desc")}

\item{arcsinhA}{(default: 0) A numeric indicating the value for 'a' in the arcsinh data transformation
equation. Should usually be 0.}

\item{arcsinhB}{(default: 0.2) A numeric indicating the value for 'b' in the arcsinh data transformation
equation. Should be 1/5 for cytof data, and 1/150 for flow data.}

\item{arcsinhC}{(default: 0) A numeric indicating the value for 'c' in the arcsinh data transformation
equation. Should usually be 0.}

\item{verbose}{A boolean specifying whether to display processing messages (default: TRUE)}
}
\value{
An S3 object of class \code{discovrExperiment}
}
\description{
Load in a dataset and perform phenograph clustering
}
\author{
Mario G Rosasco, \email{mrosasco@benaroyaresearch.org}, Virginia Muir
}
