% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringPhase.R
\name{setupDiscovrExperiment}
\alias{setupDiscovrExperiment}
\title{Prepare a dataset for DISCOV-R analysis}
\usage{
setupDiscovrExperiment(
  markerInfoFile,
  fcsInfoFile,
  parentPopulation,
  markerCommonField = "fixed",
  markerFcsField = "desc",
  arcsinhA = 0,
  arcsinhB = 0.2,
  arcsinhC = 0,
  verbose = TRUE,
  checkMemory = TRUE
)
}
\arguments{
\item{markerInfoFile}{A character string indicating the path to a .csv file. This file is expected to have columns
named useToCluster", as well as the names specified in the \code{markerCommonField} and \code{markerFcsField} variables.
Details of the \code{markerCommonField} and \code{markerFcsField} arguments are provided below; the "useToCluster"
column should have only TRUE or FALSE values. Markers with a TRUE value in this column will be used for clustering,
whereas the others will not.}

\item{fcsInfoFile}{A character string indicating the path to a file containing columns named "subject",
"cellSubset", and "filename". The "filename" field must contain paths to the .fcs files that will be used in analysis.}

\item{parentPopulation}{A character sting indicating the name of the parent population subset. Must match
one of the values in the fcsInfoFile 'cellSubset' column.}

\item{markerCommonField}{(default: "fixed") A character string indicating the
name of a column containing common marker names for human use, like "CD45"}

\item{markerFcsField}{(default: "desc") A character string indicating the
name ofa column containing the marker names in the .fcs files,like "89Y_CD45"}

\item{arcsinhA}{(default: 0) A numeric indicating the value for 'a' in the
arcsinh data transformation equation. Should usually be 0.}

\item{arcsinhB}{(default: 0.2) A numeric indicating the value for 'b' in the arcsinh
data transformation equation. Should be 1/5 for cytof data, and 1/150 for flow data.}

\item{arcsinhC}{(default: 0) A numeric indicating the value for 'c' in the
arcsinh data transformation equation. Should usually be 0.}

\item{verbose}{(default: TRUE) A boolean specifying whether to display processing messages}

\item{checkMemory}{(default: TRUE) A boolean indicating whether to check how much system memory
is available before loading the dataset. If TRUE, this function will display a message and prevent
data loading when the files take up more than 80% of the available system memory.}
}
\value{
An S3 object of class \code{discovrExperiment}
}
\description{
Loads in the marker and .fcs information and prepares a \code{discovrExperiment}
object for analysis. Default parameters values are based on the original DISCOV-R
analysis published in Wiedeman et al 2020.
}
\seealso{
\code{\link{discovrExperiment}}
}
\author{
Mario G Rosasco, \email{mrosasco@benaroyaresearch.org}, Virginia Muir
}
