% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportedUtils.R
\name{testNormalizationMethodByMarker}
\alias{testNormalizationMethodByMarker}
\title{Test a range of normalization options on a marker in a discovrExperiment}
\usage{
testNormalizationMethodByMarker(
  experiment,
  markers,
  normalizationMethod = c("none", "zScore", "warpSet"),
  seed = 12345
)
}
\arguments{
\item{experiment}{A discovrExperiment created using
\code{clusterDiscovrExperiment}, \code{normalizeDiscovrExperiment}, or
\code{metaclusterDiscovrExperiment}. In order to return metacluster numbers
for each cell, its status must be "metaclustered".}

\item{markers}{a character vector indicating the markers on which to test
normalization}

\item{normalizationMethod}{(default: c("none", "zScore", "warpSet")) a
character vector with the normalization methods to be tested. All values
should be acceptable inputs for the argument 'normalizationMethod' in
\code{normalizeDiscovrExperiment}. If \code{seed} has length > 1, this
must be a single character string, either "warpSet" or "warpSet[#]".}

\item{seed}{(default: 12345) Numeric, the random number seed(s) for warpSet
normalization, passed to \code{normalizeWarpSetMergedExpr}. If testing
various random number seeds with warpSet normalization, this should be a
numeric vector of length > 1, and \code{normalizationMethod} should be
either "warpSet" or "warpSet[#]"}
}
\value{
A reduced \code{discovrExperiment} object (or named list of objects),
each containing only the marker listed, with normalized values in
\code{experiment$mergedExprNormalizedScaled}. If \code{markers} has length >
1, the function will return a named list, with each element containing a
\code{discovrExperiment} object normalized using the given method.
}
\description{
This function tests different normalization methods on a marker in a
\code{discovrExperiment} object. Its purpose is to facilitate selection of
the appropriate normalization method for markers where this is not
immediately apparent. It can be run on a set of markers and a set of 
different normalization methods to be tested on each marker. Alternatively,
it can be run with a single warpSet normalization method and a set of
different random number seeds, to evaluate the effect of the random number
seed on the warpSet normalization results. The results can be examined for
appropriate distributions across samples, or plotted sequentially using
\code{plotDensityNormalizedExprsDiscovrExperiment}
}
\author{
Matthew J Dufort, \email{mdufort@benaroyaresearch.org}
}
